+++ src/component/index.ts
@@ -2,12 +2,15 @@ export { defineComponent } from './defineComponent'
 export { defineAsyncComponent } from './defineAsyncComponent'
 export {
   SetupFunction,
-  SetupContext,
   ComputedOptions,
   MethodOptions,
   ComponentPropsOptions,
 } from './componentOptions'
-export { ComponentInstance, ComponentRenderProxy } from './componentProxy'
+export {
+  ComponentInstance,
+  ComponentPublicInstance,
+  ComponentRenderProxy,
+} from './componentProxy'
 export { Data } from './common'
 export {
   PropType,